makes app depend on model directly
