+++ src/app.rs
@@ -1,11 +1,11 @@
 mod game_component;
 mod game_resource;
 mod particle_bundle;
-mod setup;
+mod startup;
 mod update;
 
 use game_component::GameComponent;
 pub use game_resource::GameResource;
 use particle_bundle::ParticleBundle;
-pub use setup::make_setup;
-pub use update::make_update_system;
+pub use startup::startup;
+pub use update::update;
