+++ src/app/particle_bundle.rs
@@ -1,13 +1,12 @@
 use bevy::prelude::*;
 use bevy::sprite::{ColorMaterial, Mesh2dHandle};
 
+use crate::Particle;
+
 use super::GameComponent;
 
 #[derive(Bundle)]
-pub struct ParticleBundle<P>
-where
-    P: Send + Sync + 'static,
-{
+pub struct ParticleBundle {
     pub mesh: Mesh2dHandle,
     pub material: Handle<ColorMaterial>,
     pub transform: Transform,
@@ -15,5 +14,5 @@ where
     pub visibility: Visibility,
     pub inherited_visibility: InheritedVisibility,
     pub view_visibility: ViewVisibility,
-    pub particle: GameComponent<P>,
+    pub particle: GameComponent<Particle>,
 }
