+++ src/app/update.rs
@@ -1,19 +1,21 @@
-use super::{GameComponent, GameResource};
-use bevy::prelude::{Entity, Query, Res, Transform, Vec3};
+use bevy::prelude::*;
 
-pub fn make_update_system<P, G>(
-    step: impl Fn(&P, f64, &G) -> P,
-    unwrap: impl Fn(&P) -> Vec3,
-) -> impl Fn(Query<(Entity, &mut Transform, &mut GameComponent<P>)>, Res<GameResource<G>>)
-where
-    P: Send + Sync + 'static,
-    G: Send + Sync + 'static,
-{
-    move |mut particles, global_resource| {
-        let g = &global_resource.into_inner().val;
-        for (_, mut transform, mut p) in &mut particles {
-            p.val = step(&p.val, 0.1, g);
-            transform.translation = unwrap(&p.val);
-        }
+use super::{GameComponent, GameResource};
+
+use crate::{step, Globals, Particle};
+
+pub fn update(
+    mut particles: Query<(Entity, &mut Transform, &mut GameComponent<Particle>)>,
+    mut globals: Res<GameResource<Globals>>,
+) {
+    let g = &globals.into_inner().val;
+    for (_, mut transform, mut p) in &mut particles {
+        p.val = step(&p.val, 0.1, g);
+        transform.translation = unwrap(&p.val);
     }
+}
+
+fn unwrap(p: &Particle) -> Vec3 {
+    let x = &p.position;
+    Vec3::new(x[0] as f32, x[1] as f32, 0.0)
 }
