+++ src/main.rs
@@ -3,18 +3,13 @@ mod model;
 
 use bevy::prelude::*;
 
-use app::{make_setup, make_update_system};
+use app::{startup, update};
 use model::{generate_particle, step, Globals, Particle};
 
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
-        .add_systems(Startup, make_setup::<Particle, Globals>(generate_particle))
-        .add_systems(Update, make_update_system(step, unwrap))
+        .add_systems(Startup, startup)
+        .add_systems(Update, update)
         .run();
-}
-
-fn unwrap(p: &Particle) -> Vec3 {
-    let x = &p.position;
-    Vec3::new(x[0] as f32, x[1] as f32, 0.0)
 }
